--- boat-core/src/main/kotlin/xyz/srclab/common/test/TestKit.kt
@@ -1,12 +1,0 @@
-@file:JvmName("TestKit")
-@file:JvmMultifileClass
-
-package xyz.srclab.common.test
-
-import xyz.srclab.common.collect.asToList
-
-fun equalsIgnoreOrder(a: Iterable<*>, b: Iterable<*>): Boolean {
-    val ca = a.asToList()
-    val cb = b.asToList()
-    return ca.containsAll(cb) && cb.containsAll(ca)
-}
